+++ spec/support/rspec_config/reset.rb
@@ -1,4 +1,6 @@
-RSpec.configure do |config|
+# frozen_string_literal: true
+
+RSpec.configure do |_config|
   # Because this gem uses singletons, and the tests run in a single process,
   #   the state of the Singletons needs to be reset between tests.
   UndriveGoogle::Options.instance.send(:load_yaml, true)