+++ spec/undrive_google/cli_spec.rb
@@ -14,7 +14,7 @@ RSpec.describe UndriveGoogle::CLI do
 
     include_context "with google session" do
       before do
-        file_types = (UndriveGoogle::FILE_TYPES - [:zip])
+        file_types = (UndriveGoogle::DL_FILE_TYPES - [:zip])
         file_types.each do |type|
           allow(double_file).to receive(:export_as_file).with("/My_Title.#{type}", type)
         end