+++ CHANGELOG.md
@@ -10,8 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
+## 1.0.3 - 2022-11-01
+### Added
+- Support for sweep option
+### Fixed
+- Documentation for how to use CLI (with config file, Gemfile, etc. See README)
+
 ## 1.0.2 - 2022-10-27
-###
+### Fixed
 - Fixed typo in warning when wrong version of google_drive is loaded
 
 ## 1.0.1 - 2022-10-27